requirements.txt
arrow==0.15.2
asn1crypto==0.24.0
astroid==1.6.6
aws-lambda-builders==0.4.0
aws-sam-cli==0.21.0
aws-sam-translator==1.11.0
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.7.0.1
binaryornot==0.4.4
boto3==1.9.237
botocore==1.12.237
certifi==2019.9.11
chardet==3.0.4
chevron==0.13.1
Click==7.0
CodeIntel==0.9.3
configparser==4.0.2
cookiecutter==1.6.0
cryptography==2.1.4
dateparser==0.7.2
Django==1.11.25
django-crispy-forms==1.7.2
django-filter==2.0.0
docker==4.0.2
docutils==0.15.2
enum34==1.1.6
Flask==1.0.4
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
idna==2.6
ikpdb==1.2.3
Inflector==2.0.12
ipaddress==1.0.17
isort==4.3.21
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.1.1
pathlib2==2.3.5
pbr==5.4.3
poyo==0.5.0
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.5
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
s3transfer==0.2.1
scandir==1.10.0
SecretStorage==2.3.1
serverlessrepo==0.1.9
singledispatch==3.4.0.3
six==1.12.0
stevedore==1.31.0
subprocess32==3.5.4
tzlocal==2.0.0
urllib3==1.25.6
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
websocket-client==0.56.0
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.11.2
zope.cachedescriptors==4.3.1
